cli: de-nest backend module and delete obsolete clippy warnings
The nesting was an artifact of `backend` being its own crate in the past.

Let's simplify. Also delete clippy annotations that do not apply anymore.
