+++ cli/src/api/check.rs
