+++ cli/src/api/client.rs
