+++ cli/src/api/errors.rs
