+++ cli/src/api/extension_publish.rs
