+++ cli/src/api/graphql/types/mutations/extension_publish.rs
