+++ cli/src/api/graphql/types/mutations/submit_trusted_documents.rs
