+++ cli/src/api/graphql/types/queries/fetch_subgraph_schema.rs
