+++ cli/src/api/graphql/types/queries/list_branches.rs
