+++ cli/src/api/graphql/types/queries/list_subgraphs.rs
