+++ cli/src/api/logout.rs
