+++ cli/src/api/types.rs
@@ -21,7 +21,7 @@ pub(crate) struct ProjectMetadata {
     graph_id: String,
 }
 
-#[allow(clippy::to_string_trait_impl)]
+#[expect(clippy::to_string_trait_impl)]
 impl ToString for ProjectMetadata {
     fn to_string(&self) -> String {
         serde_json::to_string(&self).expect("must parse")
