+++ cli/src/logout.rs
@@ -1,4 +1,4 @@
-use crate::{backend::api::logout, errors::CliError, output::report};
+use crate::{api::logout, errors::CliError, output::report};
 
 pub fn logout() -> Result<(), CliError> {
     logout::logout().map_err(CliError::BackendApiError)?;
